Avoid switching slide on character entry in the URL box
For some reason, it seems that the event handler here gets called even
though the URL box is supposed to catch these events and
`stopPropagation` on the event. I suspect this is due to the
interactions between React's synthetic events and the handler here,
which is added directly to the `document`. Apparently, the
`cancelBubble` property gets correctly set on the events, so we can
use that to short-circuit the handler here when it shouldn't run.